[SQL] SINGLE_VALUE panics on multiple inputs
